requirements.txt
scikit-learn==1.2.2
pandas==2.0.3
nltk==3.8.1
psycopg2-binary==2.9.7
requests==2.31.0
Pillow==10.0.0
joblib==1.3.2
spacy==3.6.1
geopy==2.3.0
streamlit-folium==0.13.0
folium==0.14.0
geocoder==1.38.1
